+++ semcore/select/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 ### Fixed
 
 - set `aria-activedescendant` for Trigger only if there is available option with current highlightedIndex.
+- The first option is highlighted, event if it is disabled.
 
 ## [4.48.6] - 2024-11-29